+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 _Released TDB_
 
+- **Breaking change**: An auth key is now provided as a `Bearer` token in the
+  HTTP `Authorization` header
 - Update minimum Python version to 3.12
 - Update Docker image to `python:3.12-slim`